Desarrollo_de_proyectos_con_IA/proyecto final/pyproject.toml
[tool.poetry]
name = "proyecto-final"
version = "0.1.0"
description = ""
authors = ["Leomorya <leomorya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
azure-storage-blob = "^12.23.0"
azure-identity = "^1.18.0"
langchain-pinecone = "^0.2.0"
langchain-community = "^0.3.1"
langchain = "^0.3.1"
tiktoken = "^0.7.0"
sentence-transformers = "^3.1.1"
torch = "^2.4.1"
streamlit = "^1.38.0"
langchain-openai = "^0.2.1"
python-dotenv = "^1.0.1"
langchain-ollama = "^0.2.0"
ftfy = "^6.2.3"
ipywidgets = "^8.1.5"
jupyter = "^1.1.1"
pypdf = "^5.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
